make scroll top button global
