+++ website/frontend/styles/index.scss
@@ -143,3 +143,36 @@ a:hover {
 img {
   pointer-events: none;
 }
+
+.scroll-top {
+  position: fixed;
+  bottom: 20px;
+  right: 20px;
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+  background-color: #145dff;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+  z-index: 1000;
+  transition: all 0.3s ease-in-out;
+  animation: scroll-top-icon 2s ease-in-out infinite;
+
+  &:hover {
+    background-color: #0e4dff;
+  }
+}
+
+@keyframes scroll-top-icon {
+  0% {
+    transform: translateY(0);
+  }
+  50% {
+    transform: translateY(-10px);
+  }
+  100% {
+    transform: translateY(0);
+  }
+}
